feat(meter): csv export
